fixed mobile layout issues for HowURPOWorks
